+++ src/daf_gui/main.py
@@ -2,10 +2,7 @@
 Main file of the DAF GUI.
 """
 from importlib.util import find_spec
-from typing import List
 from pathlib import Path
-
-from daf.misc import instance_track as it
 
 from .tabs import *
 from .edit_window_manager import *
@@ -169,7 +166,7 @@ class Application():
         self.init_output_tab()
 
         # Analytics
-        self.init_analytics_tab()
+        self.tabman_mf.add(AnalyticsTab(self.edit_mgr,padding=(dpi_5, dpi_5)), text="Analytics")
 
         # Credits tab
         self.init_credits_tab()
@@ -378,185 +375,6 @@ class Application():
         label_logo = ttk.Label(self.tab_info, image=logo)
         label_logo.image = logo
         label_logo.pack()
-
-    def init_analytics_tab(self):
-        dpi_10 = dpi_scaled(10)
-        dpi_5 = dpi_scaled(5)
-        tab_analytics = ttk.Notebook(self.tabman_mf, padding=(dpi_5, dpi_5))  # ttk.Frame(self.tabman_mf, padding=(dpi_10, dpi_10))
-        self.tabman_mf.add(tab_analytics, text="Analytics")
-
-        def create_analytic_frame(
-            getter_history: str,
-            getter_counts: str,
-            counts_coldata: dict,
-            tab_name: str
-        ):
-            """
-            Creates a logging tab.
-
-            Parameters
-            -------------
-            getter_history: str
-                The name of the LoggerBASE method that is used to retrieve actual logs.
-            getter_counts: str
-                The name of the LoggerBASE method that is used to retrieve counts.
-            counts_coldata: dict
-                Column data for TableView used for counts.
-            tab_name: str
-                The title to write inside the tab button.
-            """
-            async def analytics_load_history():
-                logger = await self.connection.get_logger()
-
-                param_object = combo_history.combo.get()
-                param_object_params = convert_to_objects(param_object.data)
-                items = await self.connection.execute_method(
-                    it.ObjectReference.from_object(logger), getter_history, **param_object_params
-                )
-                items = convert_to_object_info(items)
-                lst_history.clear()
-                lst_history.insert(tk.END, *items)
-
-            def show_log(listbox: ListBoxScrolled):
-                selection = listbox.curselection()
-                if len(selection) == 1:
-                    object_: ObjectInfo = listbox.get()[selection[0]]
-                    self.edit_mgr.open_object_edit_window(
-                        object_.class_,
-                        listbox,
-                        old_data=object_,
-                        check_parameters=False,
-                        allow_save=False
-                    )
-                else:
-                    tkdiag.Messagebox.show_error("Select ONE item!", "Empty list!")
-
-            async def delete_logs_async(primary_keys: List[int]):
-                logger = await self.connection.get_logger()
-                await self.connection.execute_method(
-                    it.ObjectReference.from_object(logger),
-                    "delete_logs",
-                    primary_keys=primary_keys  # TODO: update on server
-                )
-
-            @gui_confirm_action()
-            def delete_logs(listbox: ListBoxScrolled):
-                selection = listbox.curselection()
-                if len(selection):
-                    all_ = listbox.get()
-                    tae.async_execute(
-                        delete_logs_async([all_[i].data["id"] for i in selection]),
-                        wait=False,
-                        pop_up=True,
-                        master=self.win_main
-                    )
-                else:
-                    tkdiag.Messagebox.show_error("Select atlest one item!", "Selection error.")
-
-            frame_message = ttk.Frame(tab_analytics, padding=(dpi_5, dpi_5))
-            tab_analytics.add(frame_message, text=tab_name)
-            frame_msg_history = ttk.Labelframe(frame_message, padding=(dpi_10, dpi_10), text="Logs", bootstyle="primary")
-            frame_msg_history.pack(fill=tk.BOTH, expand=True)
-
-            combo_history = ComboEditFrame(
-                self.edit_mgr.open_object_edit_window,
-                [ObjectInfo(getattr(daf.logging.LoggerBASE, getter_history), {})],
-                frame_msg_history,
-            )
-            combo_history.pack(fill=tk.X)
-
-            frame_msg_history_bnts = ttk.Frame(frame_msg_history)
-            frame_msg_history_bnts.pack(fill=tk.X, pady=dpi_10)
-            ttk.Button(
-                frame_msg_history_bnts,
-                text="Get logs",
-                command=lambda: tae.async_execute(analytics_load_history(), wait=False, pop_up=True, master=self.win_main)
-            ).pack(side="left", fill=tk.X)
-            ttk.Button(
-                frame_msg_history_bnts,
-                command=lambda: show_log(lst_history),
-                text="View log"
-            ).pack(side="left", fill=tk.X)
-            ttk.Button(
-                frame_msg_history_bnts,
-                command=lambda: delete_logs(lst_history),
-                text="Delete selected"
-            ).pack(side="left", fill=tk.X)
-            lst_history = ListBoxScrolled(frame_msg_history)
-            lst_history.listbox.unbind_all("<Delete>")
-            lst_history.listbox.unbind_all("<BackSpace>")
-            lst_history.listbox.bind("<Delete>", lambda e: delete_logs(lst_history))
-            lst_history.listbox.bind("<BackSpace>", lambda e: delete_logs(lst_history))
-            lst_history.pack(expand=True, fill=tk.BOTH)
-
-            # Number of messages
-            async def analytics_load_num():
-                logger = await self.connection.get_logger()
-                param_object = combo_count.combo.get()
-                parameters = convert_to_objects(param_object.data)
-                count = await self.connection.execute_method(
-                    it.ObjectReference.from_object(logger),
-                    getter_counts,
-                    **parameters
-                )
-
-                tw_num.delete_rows()
-                tw_num.insert_rows(0, count)
-                tw_num.goto_first_page()
-
-            frame_num = ttk.Labelframe(frame_message, padding=(dpi_10, dpi_10), text="Counts", bootstyle="primary")
-            combo_count = ComboEditFrame(
-                self.edit_mgr.open_object_edit_window,
-                [ObjectInfo(getattr(daf.logging.LoggerBASE, getter_counts), {})],
-                frame_num,
-            )
-            combo_count.pack(fill=tk.X)
-            tw_num = tktw.Tableview(
-                frame_num,
-                bootstyle="primary",
-                coldata=counts_coldata,
-                searchable=True,
-                paginated=True,
-                autofit=True)
-
-            ttk.Button(
-                frame_num,
-                text="Calculate",
-                command=lambda: tae.async_execute(analytics_load_num(), wait=False, pop_up=True, master=self.win_main)
-            ).pack(anchor=tk.W, pady=dpi_10)
-
-            frame_num.pack(fill=tk.BOTH, expand=True, pady=dpi_5)
-            tw_num.pack(expand=True, fill=tk.BOTH)
-
-        # Message tab
-        create_analytic_frame(
-            "analytic_get_message_log",
-            "analytic_get_num_messages",
-            [
-                {"text": "Date", "stretch": True},
-                {"text": "Number of successful", "stretch": True},
-                {"text": "Number of failed", "stretch": True},
-                {"text": "Guild snowflake", "stretch": True},
-                {"text": "Guild name", "stretch": True},
-                {"text": "Author snowflake", "stretch": True},
-                {"text": "Author name", "stretch": True},
-            ],
-            "Message tracking"
-        )
-
-        # Invite tab
-        create_analytic_frame(
-            "analytic_get_invite_log",
-            "analytic_get_num_invites",
-            [
-                {"text": "Date", "stretch": True},
-                {"text": "Count", "stretch": True},
-                {"text": "Guild snowflake", "stretch": True},
-                {"text": "Guild name", "stretch": True},
-                {"text": "Invite ID", "stretch": True},
-            ],
-            "Invite tracking"
-        )
 
     def init_optional_dep_tab(self):
         dpi_10 = dpi_scaled(10)
