+++ src/daf_gui/tabs/__init__.py
@@ -1,1 +1,2 @@
 from .schema import *
+from .analytics import *
